Change how static content is managed and created.
Change the misc functions in process.py into two classes, StaticEntry
and StaticIndex.  These two classes wrap the logic to create/delete the
static HTML files for entries and indexes respectively.

Add the new concept of always generating the HTML to a GUID-based
filename.  This GUID.html file will get symlinked to the actual location
once the entry is marked as being finished.